requirements.txt
fastmcp>=2.8.1
fuzzywuzzy>=0.18.0
humancursor>=1.1.5
live-inspect>=0.1.1
markdownify>=1.1.0
pillow>=11.2.1
psutil>=7.0.0
pyautogui>=0.9.54
pyperclip>=1.8.2
python-levenshtein>=0.27.1
requests>=2.32.3
uiautomation>=2.0.24
